feat: Update Home component to display logged in state and add log out button
The Home component has been updated to display a "Logged In Yaay!!" heading and a log out button. This change includes importing the necessary dependencies and using the `useAuthActions` hook to access the `signOut` function. The purpose of this update is to enhance the user experience by providing a clear indication of the logged-in state and allowing users to log out easily.