feat: Add create workspace mutation
This commit adds a new mutation called "create" to the workspaces module. The mutation allows users to create a new workspace by providing a name. It also generates a join code and associates the workspace with the user's ID. This change enhances the functionality of the application by enabling users to create and manage workspaces.