feat: Implement useCreateWorkspace hook
The useCreateWorkspace hook has been implemented to handle the creation of workspaces. This hook provides functions for making API requests to create a new workspace, as well as managing the state and status of the request. This change enhances the functionality of the application by enabling users to create and manage workspaces.