feat: Add CreateWorkspaceModal component
A new component called CreateWorkspaceModal has been added to handle the creation of workspaces through a modal interface. This component includes form inputs for the workspace name and a "Create" button. The modal can be opened and closed using the useCreateWorkspaceModal hook. This change enhances the functionality of the application by providing a user-friendly way to create workspaces.