feat: Add Toaster component for displaying notifications
A new component called Toaster has been added to handle the display of notifications. This component uses the next-themes and sonner packages to provide a customizable and user-friendly notification system. This change enhances the user experience by providing a visually appealing and informative way to communicate important information to the user.