feat: Add getById query for getting workspace by id
This commit adds a new query function called getById to the workspaces module. This query allows fetching a workspace by its id, but only if the user is a member of that workspace. If the user is not authenticated, an "Unauthorized" error is thrown. This change enhances the security of the application by ensuring that only authorized users can access workspace data.