feat: Add update mutation to workspaces
This commit adds an update mutation to the workspaces module. The update mutation allows users with admin privileges to update the name of a workspace. It checks if the user is authorized and if the user is an admin of the workspace before updating the name. This change enhances the functionality of the workspaces module by providing the ability to modify workspace names.