feat: Add delete mutation to workspaces
Enhance the workspaces module by adding a delete mutation. This mutation allows users with admin privileges to delete a workspace. It checks if the user is authorized and if the user is an admin of the workspace before deleting it. This change improves the functionality of the workspaces module by providing the ability to remove workspaces.