feat: Add "general" channel to new workspaces
This commit adds a default "general" channel to new workspaces created using the create mutation. The "general" channel is automatically created with the workspace and provides a default channel for users to start conversations. This change enhances the functionality of the workspaces module by ensuring that every workspace has a default channel available.