feat: Add create mutation for channel creation
This commit adds a new mutation called "create" to the channels module. The mutation allows users to create a new channel by providing the channel name and workspace ID as arguments. It also includes authorization checks to ensure that only authorized users can create channels. This change enhances the functionality of the channels feature by providing a convenient way to create channels within a workspace.